Adding Current Algorithm to figure out Solutions
The current implementation creates a board to start the algorithm on
each piece of the board to figure out if there is a solution and how
many solutions there are when placing the first queen at that point.
This currently spits out the wrong solutions as it does not correctly
add to the illegalMovesBank. WIP

Resource: https://twisty-tanker-c10.notion.site/Code-Challenge-N-Queens-92e212d2b078409d930bde8352e70f53