Starting to implement backTracking
WIP General Overview: I realized that I wasn't adding backtracking, meaning the second it
sees a potential spot a queen can take, it will put the queen there
instead of checking for other solutions where that spot was not taken.
In order to implement this I need to make illegalMovesBank a local bank
for each iteration of getSolutions.

To add on I now need to branch off in getSolutions when there is a
potential placement, meaning I will need to concat the results of
getSolutions if it takes that spot and if it doesn't, which means making
a copy illegalMovesBank so they no longer share a bank while analyzing
different potential solutions.

Updates:
- Changed naming for functions
- made a generate board function as to clean up the code