Completed Algorithm
There was an issue prior where I thought I was making a copy but a
simple new assignment of a variable does not make a deeply nested copy
so I now changed addNewIllegalMovements and the board before it is
passed into getSolutions, to make deep Nested Copies of the map and nested
array respectively, and use those for the solutionAlgorithm with that
new queen placement.

I simplified addIllegalDiag as the prior code was bug prone and not
developer friendly to read -- I was crunched under time.

Added Documentation.